Feat: Update static images to videos for wasmJs Tv mode
