+++ composeApp/src/wasmJsMain/kotlin/com/ma7moud3ly/quran/platform/VideoPlayer.wasmJs.kt
@@ -1,22 +1,39 @@
 package com.ma7moud3ly.quran.platform
 
-import androidx.compose.foundation.Image
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.viewinterop.WebElementView
 import com.ma7moud3ly.quran.model.MyVideo
-import org.jetbrains.compose.resources.painterResource
+import kotlinx.browser.document
+import org.w3c.dom.HTMLVideoElement
 
+@OptIn(ExperimentalComposeUiApi::class)
 @Composable
 actual fun VideoPlayer(
     modifier: Modifier,
     state: VideoPlayerState,
     video: () -> MyVideo
 ) {
-    Image(
-        painter = painterResource(video().thumbnail),
-        contentDescription = null,
+    val video = video()
+    WebElementView(
+        factory = {
+            (document.createElement("video")
+                    as HTMLVideoElement)
+                .apply {
+                    src = video.path
+                    controls = false
+                    autoplay = true
+                    loop = true
+                    muted = true
+                    style.width = "100%"
+                    style.height = "100%"
+                    style.objectFit = "cover"
+                    style.objectPosition = "center"
+                    style.zIndex = "-1"
+                }
+        },
         modifier = modifier,
-        contentScale = ContentScale.Crop
+        update = { it.src = video.path }
     )
 }
