+++ gradle/libs.versions.toml
@@ -8,30 +8,30 @@ android-compileSdk = "36"
 android-minSdk = "24"
 android-targetSdk = "36"
 #plugins
-agp = "8.13.0"
-kotlin = "2.2.10"
-ksp = "2.2.10-2.0.2"
-composeHotReload = "1.0.0-beta06"
-composeMultiplatform = "1.8.2"
-kotlinSerialization = "2.2.10"
+agp = "8.11.1" # https://www.jetbrains.com/help/kotlin-multiplatform-dev/multiplatform-compatibility-guide.html
+kotlin = "2.2.20"
+ksp = "2.2.20-2.0.3"
+composeHotReload = "1.0.0-beta08"
+composeMultiplatform = "1.9.0"
+kotlinSerialization = "2.2.20"
 #common
-ktorClientCore = "3.2.3"
+ktorClientCore = "3.3.0"
 media3Exoplayer = "1.8.0"
 okio = "3.16.0"
 serialization-json = "1.9.0"
 kotlinxDatetime = "0.7.1"
-androidx-lifecycle = "2.9.3"
-androidx-navigation = "2.9.0-rc01"
-koinCore = "4.1.0"
+androidx-lifecycle = "2.9.4"
+androidx-navigation = "2.9.0"
+koinCore = "4.1.1"
 coilCompose = "3.3.0"
 koinAnnotations = "2.1.0"
 adaptive = "1.1.2"
 multiplatformSettings = "1.3.0"
 
 #android
-androidx-activity = "1.10.1"
+androidx-activity = "1.11.0"
 kotlinx-coroutines = "1.10.2"
-firebaseBom = "34.2.0"
+firebaseBom = "34.3.0"
 googleServices = "4.4.3"
 firebaseCrashlytics = "3.0.6"
